bugfixes in .h/.cpp

+++ Microblaze/software/src/chu_init.cpp
@@ -1,3 +1,7 @@
+#include "../include/chu_init.h"
+#include "../include/timer_core.h"
+#include "../include/uart_core.h"
+
 TimerCore _sys_timer(get_slot_addr(BRIDGE_BASE, TIMER_SLOT));
 UartCore uart(get_slot_addr(BRIDGE_BASE, UART_SLOT));
 
@@ -5,7 +9,7 @@ unsigned long long now_us() {
     return ((unsigned long) _sys_timer.read_time());
 }
 
-unsigned long now_ms() {
+unsigned long long now_ms() {
     return ((unsigned long) _sys_timer.read_time() / 1000);
 }
